Retrieve the AWS public IP address properly:
`serf` recipe retrieve the public IP on AWS.

If the AWS instance does not have public IP address,
it will return HTTP 200 return code, but no public IP address.

The retrieving process only check whether the request will be timedout or not.

Add the process whether the response is 200 or not.